% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectLSA-package.R, R/zzz.R
\docType{package}
\name{projectLSA-package}
\alias{projectLSA-package}
\title{Latent Structure Analysis Toolkit}
\description{
Provides an interactive Shiny-based environment for performing Latent Profile
Analysis (LPA), Latent Class Analysis (LCA), Latent Trait Analysis (LTA; IRT),
Exploratory Factor Analysis (EFA), Confirmatory Factor Analysis (CFA), and
Structural Equation Modeling (SEM). All analyses are conducted through a
graphical user interface without requiring users to write code.

This file records package-level imports for roxygen2 so that NAMESPACE
will include the appropriate import directives.
}
\section{Main Function}{

The primary entry point is \code{\link{run_projectLSA}}, which launches the
full interactive application.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hdmeasure/projectLSA}
  \item Report bugs at \url{https://github.com/hdmeasure/projectLSA/issues}
}

}
\author{
\strong{Maintainer}: Hasan Djidu \email{hasandjidu@gmail.com} (\href{https://orcid.org/0000-0003-1110-6815}{ORCID})

Authors:
\itemize{
  \item Hasan Djidu \email{hasandjidu@gmail.com} (\href{https://orcid.org/0000-0003-1110-6815}{ORCID})
}

Other contributors:
\itemize{
  \item Heri Retnawati (\href{https://orcid.org/0000-0002-1792-5873}{ORCID}) [contributor]
  \item Samsul Hadi (\href{https://orcid.org/0000-0003-3437-2542}{ORCID}) [contributor]
  \item Haryanto (\href{https://orcid.org/0000-0003-3322-904X}{ORCID}) [contributor]
}

}
\keyword{internal}
