% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_projectLSA}
\alias{run_projectLSA}
\title{Launch the projectLSA Shiny Application}
\usage{
run_projectLSA()
}
\value{
Launches a Shiny application (no return value)
}
\description{
This function starts the Shiny app included in the projectLSA package.
}
\examples{
if (interactive()) {
  projectLSA::run_projectLSA()
}

}
