% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{projectLSA-package}
\alias{projectLSA}
\alias{projectLSA-package}
\title{projectLSA: package-level imports}
\description{
This file records package-level imports for roxygen2 so that NAMESPACE
will include the appropriate import directives.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
